[Housekeeping] Cleanup create-account code.
