+++ client/src/pages/Upload.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { makeStyles, createStyles, Theme } from "@material-ui/core/styles";
-import { Container, Grid, Button, TextField } from "@material-ui/core";
+import { Container, Grid } from "@material-ui/core";
 import SimpleDropzone, { CSVFile } from "../components/input/SimpleDropzone";
 import AccountView from "../components/views/AccountView";
 import { getAllAccounts_allAccounts } from "../graphql/types/getAllAccounts";
@@ -9,9 +9,7 @@ import CSVColumnSelectorView, {
   ColumnMap,
 } from "../components/views/CSVColumnSelectorView";
 import PendingFileUploadView from "../components/views/PendingFileUploadView";
-import { useMutation } from "@apollo/client";
-import Queries from "../graphql/Queries";
-import { createAccount } from "../graphql/types/createAccount";
+import CreateAccountView from "../components/views/CreateAccountView";
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
@@ -32,66 +30,6 @@ const useStyles = makeStyles((theme: Theme) =>
   })
 );
 
-function CreateAccountView(props: { onSaveComplete: () => void }) {
-  const [createAccount] = useMutation<createAccount>(Queries.CREATE_ACCOUNT);
-  const [institution, setInstitution] = useState("");
-  const [name, setName] = useState("");
-  const [number, setNumber] = useState("");
-
-  const onSave = async () => {
-    await createAccount({
-      variables: {
-        options: {
-          accountName: name,
-          accountNumber: number,
-          institution: institution,
-        },
-      },
-      refetchQueries: [{ query: Queries.GET_ALL_ACCOUNTS }],
-    });
-    props.onSaveComplete();
-  };
-
-  return (
-    <Grid container direction="row" spacing={2}>
-      <Grid item>
-        <TextField
-          value={institution}
-          onChange={(e) => setInstitution(e.target.value)}
-          label="Institution"
-          variant="outlined"
-        />
-      </Grid>
-      <Grid item>
-        <TextField
-          value={name}
-          onChange={(e) => setName(e.target.value)}
-          label="Name"
-          variant="outlined"
-        />
-      </Grid>
-      <Grid item>
-        <TextField
-          value={number}
-          onChange={(e) => setNumber(e.target.value)}
-          label="Number"
-          variant="outlined"
-        />
-      </Grid>
-      <Grid item>
-        <Button
-          variant="contained"
-          color="primary"
-          size="small"
-          onClick={onSave}
-        >
-          Save
-        </Button>
-      </Grid>
-    </Grid>
-  );
-}
-
 export default function Upload() {
   const classes = useStyles();
   const [
@@ -99,7 +37,6 @@ export default function Upload() {
     setSelectedAccount,
   ] = useState<getAllAccounts_allAccounts | null>(null);
   const [pendingFiles, setPendingFiles] = useState<CSVFile[]>([]);
-  const [showCreateAccount, setShowCreateAccount] = useState(false);
   const [columnMap, setColumnMap] = useState<ColumnMap>({
     Date: 0,
     Description: 0,
@@ -120,25 +57,9 @@ export default function Upload() {
         <Grid item>
           <AccountSelector onSelectAccount={setSelectedAccount} />
         </Grid>
-        {showCreateAccount ? (
-          <Grid item>
-            <CreateAccountView
-              onSaveComplete={() => setShowCreateAccount(false)}
-            />
-          </Grid>
-        ) : (
-          <Grid item>
-            <Button
-              variant="contained"
-              color="primary"
-              size="small"
-              onClick={() => setShowCreateAccount(true)}
-            >
-              New Account
-            </Button>
-          </Grid>
-        )}
-
+        <Grid item>
+          <CreateAccountView />
+        </Grid>
         <Grid item>
           <AccountView account={selectedAccount} />
         </Grid>
